requirements.txt
opencv-python>=4.5.0
requests>=2.25.0
python-dotenv>=0.15.0
pynput
google-cloud-vision>=2.0.0
numpy>=1.19.0 